chore: logout 로직 분리
